Reorganise elvi into one dir and a separate data dir
This makes the one big Makefile rather complicated, but removes the
duplication of information in the separate Makefiles.  This also removes
the distinction between "simple" and "complex" elvi in terms of where
they are located, but obviously still exists (some elvi require complex
logic to build and/or allow complex searches).