Add `wordhippo` and `wordforms` elvi
Both of these elvi are for the same site (WordHippo).  The `wordforms`
elvis is specialised for word forms (duh) like verbs, plurals, etc.

The website uses POST requests to search for data, so it's a bit
difficult to wrangle.  My surfraw elvis generator handles it fine though
:).

The translation one is even weirder.
There are two problems:

1. The URL is mostly uniform except for *some words* which get turned
   into (presumably) its hashed version.
2. The site allows specification of the direction of translation (to- or
   from-english) but the URL doesn't indicate it at all.

Yeah, so no translations from WordHippo.