Add `wordtranslate` elvis
So it turns out that WordHippo *does* distinguish between to- and
from-English translations in the URL.

Translations to English work fine in text browsers, but the other way
around is gated by recaptcha...   It also fails to work in qutebrowser
for some reason.  Dunno why.  It's okay though, WordHippo's translations
from English are actually really bad so it's not much of a loss.  YMMV.