ddg: Add -lucky= option
I'm not sure whether the design of `--query-collapse=` is good.  I
suspect that the `*` will be very common, so perhaps excluding it would
be good?

Then, to further depart from the form of the arguments for
`--collapse=`, it could be something like:

`--query-collapse=COLLAPSE`

Where `COLLAPSE` is a *single* non-list argument.

This would allow comma (",") characters to be used since there are no
list arguments.  To allow for future extension, colons (":") wouldn't be
able to be used.

In this case, should its name change?  Similar naming to `--collapse=`
but a significantly different argument form might be confusing.