Install and uninstall manual elvi
The myanimelist elvis was old and unused.

The emojipedia (OpenSearch) elvis would ideally use the OpenSearch
infrastructure set up recently, but the site seems to reject
`opensearch-discover` with HTTP 403 (but not `opensearch2elvis`.
`opensearch-discover` doesn't allow the User-Agent header to be set
(which I presume is the reason; I haven't tested it), unlike
`opensearch2elvis`.  One of two things can be done:

1. change the OpenSearch tools to allow setting the header; or
2. make `opensearch2elvis` subsume the functionality of
  `opensearch-discover`.

I'm leaning towards (2).